+++ lib/src/http/bearer_auth.dart
@@ -1,0 +1,14 @@
+import 'http_auth.dart';
+
+class BearerAuth extends HttpAuth {
+  final String prefix;
+  final String token;
+
+  BearerAuth(this.token, {this.prefix = 'Bearer '});
+
+  @override
+  String get authorization => prefix + token;
+
+  @override
+  String toString() => authorization;
+}